finish renaming; extract LinkCommonModule
