+++ paymentsheet/src/main/java/com/stripe/android/link/injection/LinkControllerModule.kt
@@ -38,7 +38,7 @@ import javax.inject.Singleton
         LinkHoldbackExposureModule::class
     ]
 )
-internal object LinkPaymentMethodLauncherModule {
+internal object LinkControllerModule {
     @Provides
     @Singleton
     fun provideAppContext(application: Application): Context = application.applicationContext
