+++ paymentsheet/src/main/java/com/stripe/android/link/injection/LinkControllerViewModelComponent.kt
@@ -2,7 +2,7 @@ package com.stripe.android.link.injection
 
 import android.app.Application
 import androidx.lifecycle.SavedStateHandle
-import com.stripe.android.link.LinkPaymentMethodLauncherViewModel
+import com.stripe.android.link.LinkControllerViewModel
 import dagger.BindsInstance
 import dagger.Component
 import javax.inject.Singleton
@@ -10,17 +10,17 @@ import javax.inject.Singleton
 @Singleton
 @Component(
     modules = [
-        LinkPaymentMethodLauncherModule::class,
+        LinkControllerModule::class,
     ]
 )
-internal interface LinkPaymentMethodLauncherViewModelComponent {
-    val viewModel: LinkPaymentMethodLauncherViewModel
+internal interface LinkControllerViewModelComponent {
+    val viewModel: LinkControllerViewModel
 
     @Component.Factory
     interface Factory {
         fun build(
             @BindsInstance application: Application,
             @BindsInstance savedStateHandle: SavedStateHandle,
-        ): LinkPaymentMethodLauncherViewModelComponent
+        ): LinkControllerViewModelComponent
     }
 }
