+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/embedded/EmbeddedLinkExtrasModule.kt
@@ -6,6 +6,7 @@ import com.stripe.android.link.gate.DefaultLinkGate
 import com.stripe.android.link.gate.LinkGate
 import com.stripe.android.link.injection.LinkAnalyticsComponent
 import com.stripe.android.link.injection.LinkComponent
+import com.stripe.android.link.injection.LinkCommonModule
 import dagger.Binds
 import dagger.Module
 
@@ -13,6 +14,9 @@ import dagger.Module
     subcomponents = [
         LinkAnalyticsComponent::class,
         LinkComponent::class,
+    ],
+    includes = [
+        LinkCommonModule::class,
     ]
 )
 internal interface EmbeddedLinkExtrasModule {
