+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/injection/PaymentSheetCommonModule.kt
@@ -21,6 +21,7 @@ import com.stripe.android.link.gate.DefaultLinkGate
 import com.stripe.android.link.gate.LinkGate
 import com.stripe.android.link.injection.LinkAnalyticsComponent
 import com.stripe.android.link.injection.LinkComponent
+import com.stripe.android.link.injection.LinkCommonModule
 import com.stripe.android.networking.PaymentAnalyticsRequestFactory
 import com.stripe.android.paymentelement.AnalyticEventCallback
 import com.stripe.android.paymentelement.ExperimentalAnalyticEventCallbackApi
@@ -63,6 +64,9 @@ import kotlin.coroutines.CoroutineContext
         LinkAnalyticsComponent::class,
         LinkComponent::class,
     ],
+    includes = [
+        LinkCommonModule::class,
+    ]
 )
 internal abstract class PaymentSheetCommonModule {
